requirements.txt
pytest==8.3.5
PyYAML==6.0.2
Jinja2==3.1.6
pydantic==2.11.4
typer==0.15.4
antlr4-python3-runtime==4.13.2